test: own the standalone-bundle contract in core instead of nine copies
The hostExternals list, the import-specifier scrape and the process.env
guard were copy-pasted across all nine standalone-artifact suites, three
of them byte-identical. They now come from core, which every one of those
packages already depends on, and each file keeps only what is its own: a
bundled SDK that must not go external, pdf's Safari and code-splitting
rules, the plugin keys.

chat gains the no-dynamic-import guard the copies had and it lacked.
