+++ packages/pdf/resources/js/standalone-artifact.test.ts
@@ -1,33 +1,26 @@
 import { readFileSync } from "node:fs";
 import path from "node:path";
 import { expect, it } from "vitest";
+import { expectStandaloneArtifact } from "@lattice-php/core/standalone-test-support";
 
-const hostExternals = ["react", "react-dom", "react/jsx-runtime", "@lattice-php/lattice/runtime"];
 const distDir = path.resolve(import.meta.dirname, "../../dist");
 
 it("dist/plugin.js bundles the pdf engine and only imports host externals", () => {
-  const artifact = readFileSync(path.join(distDir, "plugin.js"), "utf8");
-  const specifiers = [...artifact.matchAll(/^import\b[^"'\n]*(["'])([^"'\n]+)\1/gm)].map(
-    (match) => match[2],
-  );
+  // pdf.js legitimately keeps two `import(` occurrences: the runtime-expression
+  // fake-worker fallback and a CDN-wrapper template string.
+  const artifact = expectStandaloneArtifact(path.join(distDir, "plugin.js"), {
+    dynamicImport: true,
+  });
 
-  expect(specifiers.length).toBeGreaterThan(0);
   expect(artifact).not.toContain('from"pdfjs-dist"');
-  expect(artifact).not.toContain("process.env");
 
   // getTextContent() iterates a ReadableStream with `for await`, which Safari
   // cannot do — the engine must stay on the manual reader in text-cache.ts.
   expect(artifact).not.toContain(".getTextContent(");
 
-  // pdf.js legitimately keeps two `import(` occurrences: the runtime-expression
-  // fake-worker fallback and a CDN-wrapper template string. A literal specifier
-  // after `import(` would mean a code-splitting leak instead.
+  // A literal specifier after `import(` would mean a code-splitting leak.
   expect(artifact).not.toMatch(/import\(\s*["'](?!\$\{)/);
   expect(artifact.match(/import\(/g)).toHaveLength(2);
-
-  for (const specifier of specifiers) {
-    expect(hostExternals).toContain(specifier);
-  }
 });
 
 it("ships a worker artifact matching the pinned pdfjs-dist version", () => {
