+++ packages/tree/resources/js/standalone-artifact.test.ts
@@ -1,23 +1,10 @@
-import { readFileSync } from "node:fs";
 import path from "node:path";
 import { expect, it } from "vitest";
+import { expectStandaloneArtifact } from "@lattice-php/core/standalone-test-support";
 import sourcePlugin from "./plugin";
 
-const hostExternals = ["react", "react-dom", "react/jsx-runtime", "@lattice-php/lattice/runtime"];
-
 it("dist/plugin.js only imports the standalone host externals", () => {
-  const artifact = readFileSync(path.resolve(import.meta.dirname, "../../dist/plugin.js"), "utf8");
-  const specifiers = [...artifact.matchAll(/^import\b[^"'\n]*(["'])([^"'\n]+)\1/gm)].map(
-    (match) => match[2],
-  );
-
-  expect(specifiers.length).toBeGreaterThan(0);
-  expect(artifact).not.toContain("import(");
-  expect(artifact).not.toContain("process.env");
-
-  for (const specifier of specifiers) {
-    expect(hostExternals).toContain(specifier);
-  }
+  expectStandaloneArtifact(path.resolve(import.meta.dirname, "../../dist/plugin.js"));
 });
 
 // Evaluating the bundle under gate load (vitest concurrent with composer
